Update CreateWeaponObject.md
I added some examples to make it more clear for people that it's required to use the RequestWeaponAsset native before using this one. I also changed the heading param to size since it's clearly increasing the size of the weapon and not changing the heading. 

Had this set to 8.0, you could imagine what 360 "degrees" would look like ;)
https://i.imgur.com/FKPt52J.png