Update SetParticleFxNonLoopedColour.md
The current description is in my opinion, misleading. People might think the colour parameters accept RGB values on a scale from 0 to 255 (at least, I did), while that isn't the case, as the scale goes from 0 to 1. Also it says 'not networked', but I've tested this native with another player in a server using Onesync Infinity and applied it to a networked particle effect drawing on the local player ped, and I could see the particle drawing on the other player's ped and changing colour just fine.