Update TaskFollowPointRoute.md (#943)
* Update TaskFollowPointRoute.md

Added flags and updated the name of the intunknown

* Update TaskFollowPointRoute.md

Made requested changes...hope this follows the pattern needed

* Update TaskFollowPointRoute.md

(Hopefully) fixed enum so that it's formatted properly!
Used the above example...as an example

* Update TaskFollowPointRoute.md

Committing to change to the last parameter name pending any additional enum changes needed

* Update TaskFollowPointRoute.md

Hopefully this is the appropriate change.  Still willing to try again if not. Thanks!

* Update TaskFollowPointRoute.md

Clean up old remnants and port over usage example.

* Remove extra headers

* Update TASK/TaskFollowPointRoute.md

---------

Co-authored-by: ammonia-cfx <38232208+4mmonium@users.noreply.github.com>
Co-authored-by: Dillon Skaggs <dillskaggs@gmail.com>